Added a typecast to make it a bit nicer :D

+++ src/nl/giantit/minecraft/GiantShop/core/Eco/Engines/AEco_Engine.java
@@ -65,7 +65,7 @@ public class AEco_Engine implements iEco {
 	
 	@Override
 	public double getBalance(String player) {
-		return eco.cash(player);
+		return (double) eco.cash(player);
 	}
 	
 	@Override
